refact: search.ts any타입 명시 및 타입 정리
